[refactor] 요청 처리 메시지 형식 통일 및 dto 패키징
