+++ src/main/java/com/sparta/seoulmate/controller/UserController.java
@@ -1,6 +1,9 @@
 package com.sparta.seoulmate.controller;
 
 import com.sparta.seoulmate.dto.*;
+import com.sparta.seoulmate.dto.interest.CategoryResponseDto;
+import com.sparta.seoulmate.dto.interest.InterestListResponseDto;
+import com.sparta.seoulmate.dto.user.*;
 import com.sparta.seoulmate.security.UserDetailsImpl;
 import com.sparta.seoulmate.service.EmailService;
 import com.sparta.seoulmate.service.InterestService;
@@ -89,7 +92,7 @@ public class UserController {
     @GetMapping("/signup/mail")
     public ResponseEntity mailVerification(@RequestParam String email, @RequestParam String code) {
         emailService.mailVerification(email, code);
-        return ResponseEntity.ok().body("이메일이 인증되었습니다.");
+        return ResponseEntity.ok().body("이메일 인증 완료");
     }
 
     // 인증 문자 전송
@@ -102,7 +105,7 @@ public class UserController {
     @GetMapping("/signup/sms")
     public ResponseEntity smsVerification(@RequestParam String phone, @RequestParam String code) {
         smsService.smsVerification(phone, code);
-        return ResponseEntity.ok().body("전화번호 인증이 완료되었습니다.");
+        return ResponseEntity.ok().body("전화번호 인증 성공");
     }
 
     //Blacklist를 활용한 로그아웃
@@ -126,7 +129,7 @@ public class UserController {
         }
 
         userService.updateNickname(requestDto, userDetails.getUser());
-        return ResponseEntity.ok().body(new ApiResponseDto("닉네임 변경이 완료되었습니다.", HttpStatus.OK.value()));
+        return ResponseEntity.ok().body(new ApiResponseDto("닉네임 변경 완료", HttpStatus.OK.value()));
     }
 
     // 프로필 수정(주소)
@@ -142,7 +145,7 @@ public class UserController {
         }
 
         userService.updateAddress(requestDto, userDetails.getUser());
-        return ResponseEntity.ok().body(new ApiResponseDto("주소 변경이 완료되었습니다.", HttpStatus.OK.value()));
+        return ResponseEntity.ok().body(new ApiResponseDto("주소 변경 완료", HttpStatus.OK.value()));
     }
 
     // 프로필 수정(이미지)
@@ -158,7 +161,7 @@ public class UserController {
         }
 
         userService.updateImage(file, userDetails.getUser());
-        return ResponseEntity.ok().body(new ApiResponseDto("이미지 변경이 완료되었습니다.", HttpStatus.OK.value()));
+        return ResponseEntity.ok().body(new ApiResponseDto("이미지 변경 완료", HttpStatus.OK.value()));
     }
 
     //관심사 등록 할때 쓰일 카테고리 테스트 코드
