+++ src/main/java/com/sparta/seoulmate/dto/user/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.seoulmate.dto;
+package com.sparta.seoulmate.dto.user;
 
 import com.sparta.seoulmate.entity.User;
 import com.sparta.seoulmate.entity.UserGenderEnum;
