+++ src/main/java/com/sparta/seoulmate/dto/user/UpdateNicknameRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.seoulmate.dto;
+package com.sparta.seoulmate.dto.user;
 
 import com.sparta.seoulmate.entity.User;
 import lombok.Getter;
