+++ src/main/java/com/sparta/seoulmate/dto/user/UpdatePasswordRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.seoulmate.dto;
+package com.sparta.seoulmate.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
