+++ src/main/java/com/sparta/seoulmate/entity/User.java
@@ -1,6 +1,5 @@
 package com.sparta.seoulmate.entity;
 
-import com.sparta.seoulmate.dto.UpdatePasswordRequestDto;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
