+++ src/main/java/com/sparta/seoulmate/service/InterestService.java
@@ -1,6 +1,6 @@
 package com.sparta.seoulmate.service;
 
-import com.sparta.seoulmate.dto.InterestListResponseDto;
+import com.sparta.seoulmate.dto.interest.InterestListResponseDto;
 import com.sparta.seoulmate.entity.User;
 import com.sparta.seoulmate.entity.UserInterest;
 import com.sparta.seoulmate.entity.UserInterestEnum;
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
 
 @Service
