+++ src/test/resources/application.properties
@@ -58,7 +58,11 @@ spring.data.redis.password=qw1021502
 spring.data.redis.port=6379
 spring.data.redis.database=0
 
+
 ### S3
+ACCESS_KEY = AKIAXG3FQRUIDIL2KIQD
+SECRET_KEY = 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
+BUCKET_NAME = nbcseoulmate
 #cloud.aws.credentials.accessKey=${cloud.aws.credentials.accessKey}
 #cloud.aws.credentials.secretKey=${cloud.aws.credentials.secretKey}
 #cloud.aws.region.static=${cloud.aws.region.static}
